// .velite/generated.ts
export interface Blog {
  title: string;
  description: string;
  slug: string;
  publishedAt: string;
  updatedAt?: string;
  tags: string[];
  author?: string;
  readingTime?: number;
  isPublished?: boolean;
  image: {
    src: string;
    width: number;
    height: number;
    blurDataURL?: string;
  };
  content: string;
  url?: string;
  toc?: Array<{
    title: string;
    url: string;
    items?: Array<{
      title: string;
      url: string;
      items?: any[];
    }>;
  }>;
}